Dataset wrapper moved to the base
