+++ setka/base/Trainer.py
@@ -1,127 +1,5 @@
-"""
-This is the base module of the package. It may also be called the
-engine. It contains the core classes that are used in the scripts
-such as:
-
-* OptimizerSwitch -- wrapper to facilitate work with optimizers.
-* DataSet -- a dataset class that Trainer needs.
-* Network -- a network class that Trainer needs.
-* Trainer -- class that performs training, validation and prediction.
-
-"""
-
-
-import numpy
-import os
-import random
-import torch
-import torch.utils.data
-import copy
 import collections
-import pickle
-
-class OptimizerSwitch():
-    '''
-    This class contains optimizer and the module that this
-    optimizer should optimize. The OptimizerSwitch tells trainer
-    if it should switch the modules to the training mode and back
-    to evaluation mode and if it should perform optimization for
-    this module's parameters.
-    '''
-
-    def __init__(self, train_module, optimizer, is_active=True, **kwargs):
-        '''
-        Constructs the OptimizerSwitch instance.
-
-        Args:
-            train_module (torch.nn.Module):
-
-            optimizer (torch.optim.Optimizer):
-
-            is_active (bool):
-
-            **kwargs:
-
-        Example:
-        ```
-        OptimizerSwitch(
-            net.params(),
-            torch.optim.Adam,
-            lr=3.0e-4,
-            is_active=False)
-        ```
-        '''
-
-        self.optimizer = optimizer(train_module.parameters(), **kwargs)
-        self.module = train_module
-        self.active = is_active
-
-
-class DataSet():
-    '''
-    Dataset class, where all the information about the dataset is
-    collected. It contains all the information about the dataset.
-    The dataset may be split into the subsets (that are usually
-    referred to as ```train```, ```valid``` and ```test```, but
-    you may make much more different modes for your needs).
-
-    The DataSet has the following methods:
-        * __init__ -- constructor, where the index collection
-            is ususally performed. You need to define it.
-            In the original constuctor the ```subset``` is set
-            to ```train```.
-
-        * getlen -- function that gets the size of the subset of the dataset.
-            This function gets as argument the subset ID (hashable).
-            You need to define this function in your class.
-
-        * getitem -- function that retrieves the item form the dataset.
-            This function gets as arguments the subset ID (hashable) and the
-            ID of the element in the subset (index).
-            You need to define this function in your class.
-
-
-        * getitem -- function that selects
-
-        * __len__ -- function that is called when the ```len()```
-            operator is called and returns the volume of the
-            current subset. Predefined, you do not need to redefine it.
-
-        * __getitem__ -- function that gets the item of the dataset
-            by its index. Predefined, you do not need to redefine it.
-
-
-    '''
-
-    def __init__(self):
-        pass
-
-    def getitem(self, subset, index):
-        pass
-
-    def getlen(self, subset):
-        pass
-
-    def __getitem__(self, *args):
-        args = args[0]
-        if type(args) is tuple:
-            assert(len(args) <= 2)
-            assert(len(args) > 0)
-
-            return self.getitem(args[0], args[1])
-        else:
-            if not hasattr(self, 'subset'):
-                sliced_dataset = copy.copy(self)
-                sliced_dataset.subset = args
-                return sliced_dataset
-
-            else:
-                return self.getitem(self.subset, args)
-
-    def __len__(self):
-        assert(hasattr(self, 'subset'))
-        return self.getlen(self.subset)
-
+import numpy
 
 class Trainer():
     '''
@@ -367,67 +245,3 @@ class Trainer():
         '''
         for opt_index in range(len(flags)):
             self._optimizers[opt_index].active = flags[opt_index]
-
-
-    # def save(self, name='./checkpoint'):
-    #
-    #     '''
-    #     Saves trainer to the checkpoint (stored in checkpoints directory).
-    #
-    #     Args:
-    #         name (str): name of the file where the model is saved.
-    #     '''
-    #
-    #     checkpoint = {
-    #         "epoch": self._epoch,
-    #         "iteration": self._iteration,
-    #         "model_state": self._model.module.cpu().state_dict()}
-    #
-    #     if hasattr(self, '_metrics'):
-    #         checkpoint['metrics'] = self._metrics
-    #
-    #     checkpoint['optimizers_states'] = self.get_optimizers_states()
-    #     checkpoint['optimizers_flags'] = self.get_optimizers_flags()
-    #
-    #     # for opt_index in range(len(self._optimizers)):
-    #     #     checkpoint['optimizer_state_' + str(opt_index)] = (
-    #     #         self._optimizers[opt_index].optimizer.state_dict())
-    #     #     checkpoint['optimizer_switch_' + str(opt_index)] = (
-    #     #         self._optimizers[opt_index].active)
-    #
-    #     torch.save(checkpoint,
-    #                name)
-    #
-    #     if torch.cuda.is_available():
-    #         self._model.module.cuda()
-    #
-    #
-    # def load(self, checkpoint_name):
-    #     '''
-    #     Loads model parameters from the checkpoint.
-    #
-    #     Args:
-    #         checkpoint_name (str): path to the checkpoint
-    #             to load.
-    #     '''
-    #     checkpoint = torch.load(open(checkpoint_name, 'rb'))
-    #     print("Model restored from", checkpoint_name)
-    #
-    #     self._epoch = checkpoint['epoch']
-    #     self._iteration = checkpoint['iteration']
-    #     self._model.module.load_state_dict(checkpoint["model_state"])
-    #     if 'metrics' in checkpoint:
-    #         self._metrics = checkpoint['metrics']
-    #
-    #     self.set_optimizers_flags(checkpoint['optimizers_flags'])
-    #     self.set_optimizers_states(checkpoint['optimizers_states'])
-    #     # if not new_optimizer:
-    #     # for opt_index in range(len(self._optimizers)):
-    #     #     try:
-    #     #         self._optimizers[opt_index].optimizer.load_state_dict(
-    #     #             checkpoint["optimizer_state_" + str(opt_index)])
-    #     #         self._optimizers[opt_index].active = checkpoint[
-    #     #             "optimizer_switch_" + str(opt_index)]
-    #     #     except:
-    #     #         print('Failed to load optimizer ' +
-    #     #             str(opt_index) + '.')
