+++ setka/base/__init__.py
@@ -1,0 +1,4 @@
+from .DataSet import DataSet
+from .DataSetWrapper import DataSetWrapper
+from .OptimizerSwitch import OptimizerSwitch
+from .Trainer import Trainer
